Removed floating-point ops for bit duration
before:                                 after:
AVR Memory Usage                        AVR Memory Usage
----------------                        ----------------
Device: attiny85                        Device: attiny85

Program:    6264 bytes (76.5% Full)     Program:    2906 bytes (35.5% Full)
(.text + .data + .bootloader)           (.text + .data + .bootloader)

Data:        352 bytes (68.8% Full)     Data:         88 bytes (17.2% Full)
(.data + .bss + .noinit)                (.data + .bss + .noinit)

Holy.  Shit.  One "(float)"* cost 3358 bytes (40.9%) of program space and 264
bytes (51.6%) of SRAM.

* the initial_timer0_seed change was just stylistic

+++ main/src/usi_serial.c
@@ -100,11 +100,10 @@ void usi_serial_init(const USISerialRegisters *_reg,
     */
     
     // cycles/bit = F_CPU/(baud * prescale)
-    timer0_seed = (uint8_t)(( F_CPU / ((float) baud_rate)) / 8);
+    timer0_seed = (F_CPU / baud_rate) / 8;
     
     // 1.5 times timer0_seed
-    initial_timer0_seed =
-        ((uint8_t)(( timer0_seed * 3 ) / 2)) - PCINT_STARTUP_DELAY;
+    initial_timer0_seed = ((timer0_seed * 3) / 2) - PCINT_STARTUP_DELAY;
 
     rxState = USIRX_STATE_IDLE;
     txState = USITX_STATE_IDLE;
